+++ .env.template
@@ -1,9 +1,15 @@
+
 # .env
 # This file contains environment variables required for the application.
 # fill in the necessary API keys and settings.
 # Please refer to the respective API documentation for more details on obtaining API keys.
 
 MAIN_SYSTEM_PROMPT="You are an AI Assistant that can answer questions and perform complex tasks using the available tools and by asking available experts questions. You are designed to assist users by using the tools available to you to gather data and complete actions required to best complete the users' request. Before providing a final response, take the time to reason and work out the best possible response for the given user request."
+
+LOGGING_ENABLED=False
+LOGGING_LEVEL=Debug
+LOGGING_FILE=logs/ksGPT.log
+LOGGING_FORMAT=%(name)s - %(levelname)s - %(message)s
 
 ##############################################################################################################
 # TTS SETTINGS
@@ -23,12 +29,12 @@ TTS_RATE=150
 
 ##############################################################################################################
 
-ENABLE_GMAILTOOLSPLUGIN=False
-ENABLE_NHTSAVPICPLUGIN=True
 ENABLE_ACCUWEATHERPLUGIN=False
 ENABLE_GEMINIPROPLUGIN=False
-ENABLE_GOOGLESEARCHPLUGIN=Fasle
-ENABLE_NEWSEXPERTPLUGIN=False
+ENABLE_GMAILPLUGIN=False
+ENABLE_GOOGLESEARCHPLUGIN=False
+ENABLE_NEWSPLUGIN=False
+ENABLE_NHTSAVPICPLUGIN=True
 ENABLE_SYSTEMCOMMANDSPLUGIN=False
 
 
@@ -100,6 +106,9 @@ GOOGLE_CSE_ID=
 # Your Gmail address to send email from (required if tools are enabled)
 
 GMAIL_ADDRESS=
+GOOGLE_REDIRECT_URI="http://localhost:8000"
+GOOGLE_CLIENT_ID=
+GOOGLE_CLIENT_SECRET=
 
 #########################################################################################
 # NEWSAPI.org API SETTINGS
