+++ .gitignore
@@ -163,3 +163,4 @@ cython_debug/
 *.ini
 *.json
 token.json
+/logs
