+++ plugins/_news_plugin/requirements.txt
