+++ plugins/template/_my_plugin/my_plugin_base.py
@@ -1,23 +1,17 @@
+
+# !/usr/bin/env python
+# coding: utf-8
 # Filename: my_plugin.py
-# Path: plugins/_my_plugin/my_plugin.py
+# Path: plugins/template/_my_plugin/my_plugin.py
 
 """
 This module defines the MyPlugin plugin.
 """
-import os
-import sys
 
-# Get the directory of the current file
-current_dir = os.path.dirname(os.path.abspath(__file__))
-
-# Add the directory to the Python path to ensure imports work
-sys.path.insert(0, current_dir)
-
-from my_tools import (
+from plugins.template._my_plugin.my_plugin_tools import (
     my_tool_list,
     available_functions as my_functions
 )
-
 from plugins.plugin_base import PluginBase
 
 
@@ -31,7 +25,7 @@ class MyPlugin(PluginBase):
         """
         self.load_plugin_tools()
 
-    def load_plugin_tools(self):
+    async def load_plugin_tools(self):
         """
         Load tools and functions from accompanying scripts.
         """
