+++ plugins/template/docs/PLUGINTEMPLATE.md
